make exptimes 32bit wide, it's spec compliant and will save some space on the expiry heap

+++ command.go
@@ -25,7 +25,7 @@ type StorageCommand struct {
   command     string
   key         string
   flags       uint32
-  exptime     uint64
+  exptime     uint32
   bytes       uint32
   cas_unique  uint64
   noreply     bool
@@ -129,9 +129,9 @@ func (sc *StorageCommand) parse(line []string) os.Error {
   sc.key = line[1]
   sc.flags = uint32(flags)
   if exptime < secondsInMonth {
-    sc.exptime = time.Seconds() + exptime;
+    sc.exptime = uint32(time.Seconds()) + uint32(exptime);
   } else {
-    sc.exptime = exptime
+    sc.exptime = uint32(exptime)
   }
   sc.bytes = uint32(bytes)
   sc.cas_unique = casuniq
